+++ app/Http/Controllers/Gallery/AlbumListing/AlbumChildrenController.php
@@ -8,23 +8,23 @@
 
 namespace App\Http\Controllers\Gallery\AlbumListing;
 
-use App\Constants\AccessPermissionConstants as APC;
+use App\Actions\Album\StructOfArrays\QueryChildrenForAlbum;
+use App\Actions\Album\StructOfArrays\QueryChildrenForPerson;
+use App\Actions\Album\StructOfArrays\QueryChildrenForTag;
+use App\Actions\Album\StructOfArrays\QueryRightsForAlbum;
+use App\Actions\Album\StructOfArrays\QueryRightsForMatchingAlbums;
 use App\Contracts\Models\AbstractAlbum;
-use App\DTO\AlbumSortingCriterion;
 use App\Enum\ColumnSortingType;
 use App\Enum\OrderSortingType;
 use App\Enum\TimelineAlbumGranularity;
 use App\Enum\TitleBucketMode;
-use App\Http\Controllers\Gallery\AlbumListController;
 use App\Http\Requests\Album\GetAlbumBucketsRequest;
 use App\Http\Requests\Album\GetAlbumChildrenDataRequest;
 use App\Http\Requests\Album\GetAlbumChildrenRightsRequest;
 use App\Http\Resources\V3\AlbumBucketResource;
-use App\Http\Resources\V3\AlbumChildrenDataResource;
-use App\Http\Resources\V3\AlbumChildrenRightsResource;
-use App\Models\AccessPermission;
+use App\Http\Resources\V3\AlbumDataResource;
+use App\Http\Resources\V3\AlbumRightsResource;
 use App\Models\Album;
-use App\Models\Extensions\SortingDecorator;
 use App\Models\PersonAlbum;
 use App\Models\TagAlbum;
 use App\Models\User;
@@ -35,11 +35,8 @@ use App\Repositories\ConfigManager;
 use App\Services\AlbumBucketComputer;
 use App\Services\Cache\CacheKeyProvider;
 use App\Services\Cache\ManagedCacheService;
-use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Routing\Controller;
-use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Facades\DB;
 use function Safe\mktime;
 
 /**
@@ -60,6 +57,11 @@ class AlbumChildrenController extends Controller
 		protected ManagedCacheService $managed_cache_service,
 		protected CacheKeyProvider $cache_key_provider,
 		protected ConfigManager $config_manager,
+		protected QueryChildrenForAlbum $query_children_for_album,
+		protected QueryChildrenForTag $query_children_for_tag,
+		protected QueryChildrenForPerson $query_children_for_person,
+		protected QueryRightsForAlbum $query_rights_for_album,
+		protected QueryRightsForMatchingAlbums $query_rights_for_matching_albums,
 	) {
 	}
 
@@ -175,7 +177,7 @@ class AlbumChildrenController extends Controller
 
 	// ── GET /Albums/{album_id} ──────────────────────────────────────
 
-	public function index(GetAlbumChildrenDataRequest $request): AlbumChildrenDataResource
+	public function index(GetAlbumChildrenDataRequest $request): AlbumDataResource
 	{
 		$album = $request->album();
 		/** @var User|null $user */
@@ -205,193 +207,28 @@ class AlbumChildrenController extends Controller
 				$this->cache_key_provider->albumChildrenTag($album->get_id()),
 				$this->cache_key_provider->userTag($user?->id),
 			],
-			fn (): AlbumChildrenDataResource => $this->queryChildren($album, $user),
+			fn (): AlbumDataResource => $this->queryChildren($album, $user),
 			ttl: $ttl,
 		);
 	}
 
-	private function queryChildren(AbstractAlbum $album, ?User $user): AlbumChildrenDataResource
+	private function queryChildren(AbstractAlbum $album, ?User $user): AlbumDataResource
 	{
 		if ($album instanceof Album) {
-			$query = Album::query()->where('parent_id', '=', $album->id);
-			// applyVisibilityFilter() already joins computed_access_permissions
-			// internally (prepareModelQueryOrFail()) — must not join it again.
-			$query = $this->album_query_policy->applyVisibilityFilter($query, $user);
-
-			$sorting = $album->getEffectiveAlbumSorting();
-			$direction = $sorting->order === OrderSortingType::DESC ? 'desc' : 'asc';
-			// Order by bucket_id first (mirrors queryBuckets() exactly,
-			// "unknown" always last) so grouping these rows by bucket_id
-			// reproduces the buckets endpoint's own row order; this is required, not
-			// redundant with the effective-column order below, because under
-			// title_bucket_mode=date_prefix the parsed-date bucket_id and the
-			// title_base/title_index sort key are unrelated dimensions of the same
-			// string, so same-bucket rows would not otherwise stay contiguous.
-			$query->orderByRaw('(albums.bucket_id IS NULL) ASC')
-				->orderBy('albums.bucket_id', $direction);
-			(new SortingDecorator($query))->orderBy($sorting->column, $sorting->order)->applyOrdering();
-
-			return $this->fetch($query, $user);
+			return $this->query_children_for_album->do($album, $user);
 		}
 
 		if ($album instanceof TagAlbum) {
-			if (!$this->config_manager->getValueAsBool('TA_albums_listing_enabled')) {
-				return $this->toResource(collect(), $user);
-			}
-
-			$unlocked_album_ids = AlbumPolicy::getUnlockedAlbumIDs();
-			$tag_ids = $album->tags->pluck('id')->all();
-			$query = $this->album_repository->queryMatchingAlbumsForTag($tag_ids, $album->is_and, $unlocked_album_ids);
-			// queryMatchingAlbumsForTag()/applyBrowsabilityFilter() do not join
-			// computed_access_permissions on the outer query — add it here.
-			$this->album_query_policy->joinSubComputedAccessPermissions($query, 'albums.id', 'left', '', false, $user);
-			// No parent-governed bucket_id concept applies to a dynamically-matched,
-			// disparately-parented result set (tier 1 excludes these types entirely) -
-			// order by the same instance-wide default v2's paginated listing already
-			// uses for this type.
-			$default_sorting = AlbumSortingCriterion::createDefault();
-			(new SortingDecorator($query))->orderBy($default_sorting->column, $default_sorting->order)->applyOrdering();
-
-			return $this->fetch($query, $user);
-		}
-
-		// PersonAlbum — the only remaining type GetAlbumChildrenDataRequest resolves.
-		if (!$this->config_manager->getValueAsBool('PA_albums_listing_enabled')) {
-			return $this->toResource(collect(), $user);
+			return $this->query_children_for_tag->do($album, $user);
 		}
 
 		/** @var PersonAlbum $album */
-		$unlocked_album_ids = AlbumPolicy::getUnlockedAlbumIDs();
-		$query = $this->album_repository->queryMatchingAlbumsForPerson($album, $unlocked_album_ids);
-		$this->album_query_policy->joinSubComputedAccessPermissions($query, 'albums.id', 'left', '', false, $user);
-		$default_sorting = AlbumSortingCriterion::createDefault();
-		(new SortingDecorator($query))->orderBy($default_sorting->column, $default_sorting->order)->applyOrdering();
-
-		return $this->fetch($query, $user);
-	}
-
-	/**
-	 * @param Builder<Album> $query
-	 */
-	private function fetch(Builder $query, ?User $user): AlbumChildrenDataResource
-	{
-		// The album's own public/anonymous grant,
-		// independent of the requesting viewer — distinct from
-		// computed_access_permissions above, which reflects the *viewer's*
-		// effective access. A unique index on
-		// (base_album_id, user_id_unique_key, user_group_id_unique_key)
-		// guarantees at most one such row per album, so this left join can
-		// never fan out the result set. Joined via a narrow-column subquery
-		// (mirrors joinBaseAlbumOwnerId()/joinSubComputedAccessPermissions()'s
-		// own convention), not a raw table join — access_permissions carries
-		// its own created_at/updated_at columns that would otherwise collide
-		// with SortingDecorator's unqualified `ORDER BY created_at`.
-		$query->joinSub(
-			query: DB::table('access_permissions')
-				->select(['base_album_id', 'is_link_required'])
-				->whereNull('user_id')
-				->whereNull('user_group_id'),
-			as: 'public_access_permissions',
-			first: 'public_access_permissions.base_album_id',
-			operator: '=',
-			second: 'albums.id',
-			type: 'left'
-		);
-
-		$rows = $query
-			->select([
-				'albums.id',
-				'base_albums.title',
-				'albums.cover_id',
-				'albums.auto_cover_id_max_privilege',
-				'albums.auto_cover_id_least_privilege',
-				'base_albums.owner_id',
-				'albums.bucket_id',
-				'computed_access_permissions.password',
-				'base_albums.is_nsfw',
-				'base_albums.is_pinned',
-				'public_access_permissions.base_album_id as public_grant_id',
-				'public_access_permissions.is_link_required as public_is_link_required',
-				'albums.num_children',
-				'albums.num_photos',
-				'base_albums.created_at',
-				'albums.min_taken_at',
-				'albums.max_taken_at',
-			])
-			->selectRaw('SUBSTR(base_albums.description, 1, 100) as description')
-			->toBase()
-			->get();
-
-		return $this->toResource($rows, $user);
-	}
-
-	/**
-	 * @param Collection<int,object{id:string,title:string,description:?string,cover_id:?string,auto_cover_id_max_privilege:?string,auto_cover_id_least_privilege:?string,owner_id:int,bucket_id:?string,password:?string,is_nsfw:mixed,is_pinned:mixed,public_grant_id:?string,public_is_link_required:mixed,num_children:int|string,num_photos:int|string,created_at:string,min_taken_at:?string,max_taken_at:?string}> $rows
-	 */
-	private function toResource(Collection $rows, ?User $user): AlbumChildrenDataResource
-	{
-		$ids = [];
-		$titles = [];
-		$descriptions = [];
-		$cover_ids = [];
-		$bucket_ids = [];
-		$owner_ids = [];
-		$is_password_requireds = [];
-		$is_nsfws = [];
-		$is_pinneds = [];
-		$is_publics = [];
-		$is_link_requireds = [];
-		$has_subalbums = [];
-		$num_photos = [];
-		$num_subalbums = [];
-		$created_ats = [];
-		$min_taken_ats = [];
-		$max_taken_ats = [];
-
-		foreach ($rows as $row) {
-			$ids[] = $row->id;
-			$titles[] = $row->title;
-			$descriptions[] = $row->description ?? '';
-			$cover_ids[] = AlbumListController::resolveCoverId($row, $user);
-			$bucket_ids[] = $row->bucket_id ?? 'unknown';
-			$owner_ids[] = (string) $row->owner_id;
-			$is_password_requireds[] = $row->password !== null;
-			$is_nsfws[] = filter_var($row->is_nsfw, FILTER_VALIDATE_BOOLEAN);
-			$is_pinneds[] = filter_var($row->is_pinned, FILTER_VALIDATE_BOOLEAN);
-			$is_publics[] = $row->public_grant_id !== null;
-			$is_link_requireds[] = filter_var($row->public_is_link_required, FILTER_VALIDATE_BOOLEAN);
-			$has_subalbums[] = ((int) $row->num_children) > 0;
-			$num_photos[] = (int) $row->num_photos;
-			$num_subalbums[] = (int) $row->num_children;
-			$created_ats[] = $row->created_at;
-			$min_taken_ats[] = $row->min_taken_at;
-			$max_taken_ats[] = $row->max_taken_at;
-		}
-
-		return new AlbumChildrenDataResource(
-			ids: $ids,
-			titles: $titles,
-			descriptions: $descriptions,
-			cover_ids: $cover_ids,
-			bucket_ids: $bucket_ids,
-			owner_ids: $owner_ids,
-			is_password_requireds: $is_password_requireds,
-			is_nsfws: $is_nsfws,
-			is_pinneds: $is_pinneds,
-			is_publics: $is_publics,
-			is_link_requireds: $is_link_requireds,
-			has_subalbums: $has_subalbums,
-			num_photos: $num_photos,
-			num_subalbums: $num_subalbums,
-			created_ats: $created_ats,
-			min_taken_ats: $min_taken_ats,
-			max_taken_ats: $max_taken_ats,
-		);
+		return $this->query_children_for_person->do($album, $user);
 	}
 
 	// ── GET /Albums/{album_id}/rights ───────────────────────────────
 
-	public function rights(GetAlbumChildrenRightsRequest $request): AlbumChildrenRightsResource
+	public function rights(GetAlbumChildrenRightsRequest $request): AlbumRightsResource
 	{
 		$album = $request->album();
 		/** @var User|null $user */
@@ -414,179 +251,38 @@ class AlbumChildrenController extends Controller
 				$this->cache_key_provider->albumChildrenTag($album->get_id()),
 				$this->cache_key_provider->userTag($user?->id),
 			],
-			fn (): AlbumChildrenRightsResource => $this->queryRights($album, $user),
+			fn (): AlbumRightsResource => $this->queryRights($album, $user),
 			ttl: $ttl,
 		);
 	}
 
-	private function queryRights(AbstractAlbum $album, ?User $user): AlbumChildrenRightsResource
+	private function queryRights(AbstractAlbum $album, ?User $user): AlbumRightsResource
 	{
 		if ($album instanceof Album) {
-			return $this->queryRightsForAlbum($album, $user);
+			return $this->query_rights_for_album->do($album, $user);
 		}
 
 		if ($album instanceof TagAlbum) {
 			if (!$this->config_manager->getValueAsBool('TA_albums_listing_enabled')) {
-				return $this->emptyMatchingAlbumsResource($album);
+				return $this->query_rights_for_matching_albums->emptyResource($album);
 			}
 
 			$unlocked_album_ids = AlbumPolicy::getUnlockedAlbumIDs();
 			$tag_ids = $album->tags->pluck('id')->all();
 			$query = $this->album_repository->queryMatchingAlbumsForTag($tag_ids, $album->is_and, $unlocked_album_ids);
 
-			return $this->queryRightsForMatchingAlbums($album, $query, $user);
+			return $this->query_rights_for_matching_albums->do($album, $query, $user);
 		}
 
 		// PersonAlbum — the only remaining type GetAlbumChildrenRightsRequest resolves.
 		if (!$this->config_manager->getValueAsBool('PA_albums_listing_enabled')) {
-			return $this->emptyMatchingAlbumsResource($album);
+			return $this->query_rights_for_matching_albums->emptyResource($album);
 		}
 
 		/** @var PersonAlbum $album */
 		$unlocked_album_ids = AlbumPolicy::getUnlockedAlbumIDs();
 		$query = $this->album_repository->queryMatchingAlbumsForPerson($album, $unlocked_album_ids);
 
-		return $this->queryRightsForMatchingAlbums($album, $query, $user);
-	}
-
-	private function queryRightsForAlbum(Album $album, ?User $user): AlbumChildrenRightsResource
-	{
-		$is_admin = $user?->may_administrate === true;
-
-		$query = Album::query()->where('parent_id', '=', $album->id);
-		$query = $this->album_query_policy->applyVisibilityFilter($query, $user);
-
-		if ($is_admin) {
-			// Mirrors AlbumQueryPolicy::applyVisibilityFilter()/applyReachabilityFilter()'s
-			// own admin early-return — neither the grants join
-			// nor the can_delete_children exists() query ever runs.
-			$ids = $query->select(['albums.id'])->toBase()->pluck('id')->all();
-			$count = count($ids);
-
-			return new AlbumChildrenRightsResource(
-				owner_id: (string) $album->owner_id,
-				can_delete_children: true,
-				can_move_children: true,
-				ids: $ids,
-				grants_edit: array_fill(0, $count, true),
-				grants_download: array_fill(0, $count, true),
-			);
-		}
-
-		[$ids, $grants_edit, $grants_download] = $this->fetchGrants($query, $user);
-		$can_delete_children = $this->canDeleteChildren($album, $user);
-
-		return new AlbumChildrenRightsResource(
-			owner_id: (string) $album->owner_id,
-			can_delete_children: $can_delete_children,
-			can_move_children: $can_delete_children,
-			ids: $ids,
-			grants_edit: $grants_edit,
-			grants_download: $grants_download,
-		);
-	}
-
-	/**
-	 * @param Builder<Album> $query
-	 */
-	private function queryRightsForMatchingAlbums(TagAlbum|PersonAlbum $album, Builder $query, ?User $user): AlbumChildrenRightsResource
-	{
-		$is_admin = $user?->may_administrate === true;
-
-		if ($is_admin) {
-			$ids = $query->select(['albums.id'])->toBase()->pluck('id')->all();
-			$count = count($ids);
-
-			return new AlbumChildrenRightsResource(
-				owner_id: (string) $album->owner_id,
-				can_delete_children: false,
-				can_move_children: false,
-				ids: $ids,
-				grants_edit: array_fill(0, $count, true),
-				grants_download: array_fill(0, $count, true),
-			);
-		}
-
-		[$ids, $grants_edit, $grants_download] = $this->fetchGrants($query, $user);
-
-		return new AlbumChildrenRightsResource(
-			owner_id: (string) $album->owner_id,
-			// No single shared parent's access_permissions could uniformly
-			// apply to a dynamically-matched, disparately-parented set.
-			can_delete_children: false,
-			can_move_children: false,
-			ids: $ids,
-			grants_edit: $grants_edit,
-			grants_download: $grants_download,
-		);
-	}
-
-	private function emptyMatchingAlbumsResource(TagAlbum|PersonAlbum $album): AlbumChildrenRightsResource
-	{
-		return new AlbumChildrenRightsResource(
-			owner_id: (string) $album->owner_id,
-			can_delete_children: false,
-			can_move_children: false,
-			ids: [],
-			grants_edit: [],
-			grants_download: [],
-		);
-	}
-
-	/**
-	 * `getComputedAccessPermissionSubQuery(full: true, ...)` applies no
-	 * internal `GROUP BY` — a caller in multiple groups with
-	 * separate matching grants on the same child would otherwise produce
-	 * duplicate joined rows. `GROUP BY` + `MAX()` here correctly OR-merges
-	 * them: any matching group/user/public row granting a right is enough.
-	 *
-	 * @param Builder<Album> $query
-	 *
-	 * @return array{0:string[],1:bool[],2:bool[]}
-	 */
-	private function fetchGrants(Builder $query, ?User $user): array
-	{
-		$this->album_query_policy->joinSubComputedAccessPermissions($query, 'albums.id', 'left', 'grants_', true, $user);
-
-		$rows = $query
-			->select(['albums.id'])
-			->selectRaw('MAX(grants_computed_access_permissions.grants_edit) as grants_edit')
-			->selectRaw('MAX(grants_computed_access_permissions.grants_download) as grants_download')
-			->groupBy('albums.id')
-			->toBase()
-			->get();
-
-		$ids = [];
-		$grants_edit = [];
-		$grants_download = [];
-		foreach ($rows as $row) {
-			$ids[] = $row->id;
-			$grants_edit[] = filter_var($row->grants_edit, FILTER_VALIDATE_BOOLEAN);
-			$grants_download[] = filter_var($row->grants_download, FILTER_VALIDATE_BOOLEAN);
-		}
-
-		return [$ids, $grants_edit, $grants_download];
-	}
-
-	/**
-	 * Mirrors {@see \App\Policies\AlbumPolicy::canDelete()}'s parent-scoped
-	 * `AccessPermission` query verbatim, addressed directly at `$album->id`
-	 * (equivalent to `$abstract_album->parent_id` there, since every
-	 * returned child's parent *is* `$album->id`).
-	 */
-	private function canDeleteChildren(Album $album, ?User $user): bool
-	{
-		if ($user === null) {
-			return false;
-		}
-
-		return AccessPermission::query()
-			->where(APC::BASE_ALBUM_ID, '=', $album->id)
-			->where(
-				fn ($query) => $query->where(APC::USER_ID, '=', $user->id)
-					->orWhereIn(APC::USER_GROUP_ID, $user->user_groups->pluck('id'))
-			)
-			->where(APC::GRANTS_DELETE, '=', true)
-			->exists();
+		return $this->query_rights_for_matching_albums->do($album, $query, $user);
 	}
 }
