+++ app/Http/Resources/V3/AlbumCategoryRightsResource.php
@@ -14,7 +14,7 @@ use Spatie\TypeScriptTransformer\Attributes\TypeScript;
 /**
  * Response body of `GET /api/v3/Albums/tags/rights` —
  * a flat tag-album catalogue has no parent-child relationship
- * to check, so unlike {@see AlbumChildrenRightsResource} there is no
+ * to check, so unlike {@see AlbumRightsResource} there is no
  * `can_delete_children`/`can_move_children` concept, and no whole-response
  * `owner_id` (each tag album has its own, carried per-row via the sibling
  * `/Albums/tags` listing's `owner_ids[]` instead). `grants_delete[]` is
